% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_complete.R
\name{est_complete}
\alias{est_complete}
\title{Complete measure.}
\usage{
est_complete(X, D, Q, num_obs, num_dim, num_comp, index_comp)
}
\arguments{
\item{X}{A matrix.}

\item{D}{A matrix.}

\item{Q}{A scalar.}

\item{num_obs}{The number of observations.}

\item{num_dim}{The number of dimensions.}

\item{num_comp}{The number of components.}

\item{index_comp}{The index of components.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Complete measure.
}
