% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{mdm}
\alias{mdm}
\title{Mutual Dependence Measures}
\usage{
mdm(X, dim_comp = NULL, dist_comp = FALSE, type = c("comp_simp"))
}
\arguments{
\item{X}{A matrix.}

\item{dim_comp}{The dimensions of components in X \code{X}.}

\item{dist_comp}{Logical. If \code{TRUE}, the distances between all components
in X will be returned.}

\item{type}{The type of mutual dependence measures, including
- \code{asym_dcov}: asymmetric measure based on distance covariance
- \code{sym_dcov}: symmectric measure based on distance covariance
- \code{comp}: complete measure based on complete V-statistics
- \code{comp_simp}: simplified complete measure based on incomplete V-statistics
- \code{asym_comp}: asymmetric measure based on complete measure
- \code{asym_comp_simp}: simplified asymmetric measure based on simplified complete measure
- \code{sym_comp}: symmectric measure based on complete measure
- \code{sym_comp_simp}: simplified symmectric measure based on simplified complete measure}
}
\value{
\code{mdm} returns a list containing the following components:
\item{stat}{The value of mutual dependence measure.}
\item{dist}{The distances between all components.}
}
\description{
\code{mdm} computes mutual dependence measures.
}
\examples{
X <- matrix(rnorm(30), 10, 3)
mdm(X, type = 'asym_dcov')
mdm(X, type = 'sym_dcov')
mdm(X, type = 'comp')
mdm(X, type = 'comp_simp')
mdm(X, type = 'asym_comp')
mdm(X, type = 'asym_comp_simp')
mdm(X, type = 'sym_comp')
mdm(X, type = 'sym_comp_simp')
}
\references{
Jin, Z. and Matteson, D. S. (2017).
  Generalizing Distance Covariance to Measure and Test Multivariate Mutual Dependence.
  arXiv preprint arXiv:1709.02532.
  \url{https://arxiv.org/abs/1709.02532}.
}
