% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm_test.R
\name{mdm_test}
\alias{mdm_test}
\title{Mutual Independence Tests}
\usage{
mdm_test(X, dim_comp = NULL, num_perm = NULL, type = c("comp_simp"))
}
\arguments{
\item{X}{A matrix or data frame, where rows represent samples, and columns represent variables.}

\item{dim_comp}{The numbers of variables included by all components in \code{X}.
If omitted, each component is assumed to contain exactly one variable.}

\item{num_perm}{The number of permutation samples drawn to approximate the asymptotic distributions
of mutual dependence measures.}

\item{type}{The type of mutual dependence measures, including
(1) \code{asym_dcov}: asymmetric measure based on distance covariance;
(2) \code{sym_dcov}: symmectric measure based on distance covariance;
(3) \code{comp}: complete measure based on complete V-statistics;
(4) \code{comp_simp}: simplified complete measure based on incomplete V-statistics;
(5) \code{asym_comp}: asymmetric measure based on complete measure;
(6) \code{asym_comp_simp}: simplified asymmetric measure based on simplified complete measure;
(7) \code{sym_comp}: symmectric measure based on complete measure;
(8) \code{sym_comp_simp}: simplified symmectric measure based on simplified complete measure.}
}
\value{
\code{mdm_test} returns a list containing the following components:
\item{stat}{The value of mutual dependence measure.}
\item{pval}{The p-value of mutual independence test.}
}
\description{
\code{mdm_test} tests mutual independence of all components in \code{X},
where each component contains one variable (univariate) or more variables (multivariate).
All tests are implemented as permutation tests.
}
\examples{
X <- matrix(rnorm(30), 10, 3)
mdm_test(X, type = 'comp_simp')
}
\references{
Jin, Z. and Matteson, D. S. (2017).
  Generalizing Distance Covariance to Measure and Test Multivariate Mutual Dependence.
  arXiv preprint arXiv:1709.02532.
  \url{https://arxiv.org/abs/1709.02532}.
}
