% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDMeasure-package.R
\docType{package}
\name{MDMeasure-package}
\alias{MDMeasure-package}
\alias{MDMeasure}
\title{Mutual Dependence Measures via Energy Statistics}
\description{
MDMeasure: A package for mutual dependence measures via energy statistics
}
\details{
The MDMeasure package provides measures of mutual dependence and tests of mutual independence.
}
\section{Measuring mutual dependence}{

The mutual dependence measures include: \cr
- asymmetric measure \eqn{\mathcal{R}_n} based on distance covariance \eqn{\mathcal{V}_n} \cr
- symmetric measure \eqn{\mathcal{S}_n} based on distance covariance \eqn{\mathcal{V}_n} \cr
- complete measure \eqn{\mathcal{Q}_n} based on complete V-statistics \cr
- simplified complete measure \eqn{\mathcal{Q}_n^\star} based on incomplete V-statistics \cr
- asymmetric measure \eqn{\mathcal{J}_n} based on complete measure \eqn{\mathcal{Q}_n} \cr
- simplified asymmetric measure \eqn{\mathcal{J}_n^\star} based on simplified complete measure 
  \eqn{\mathcal{Q}_n^\star} \cr
- symmetric measure \eqn{\mathcal{I}_n} based on complete measure \eqn{\mathcal{Q}_n} \cr
- simplified symmetric measure \eqn{\mathcal{I}_n^\star} based on simplified complete measure 
  \eqn{\mathcal{Q}_n^\star}
}

\section{Testing mutual independence}{

The mutual independence tests based on the mutual dependence measures are implemented as permutation 
tests.
}

\author{
Ze Jin \email{zj58@cornell.edu},
  David S. Matteson \email{matteson@cornell.edu}
}
