% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{mdm}
\alias{mdm}
\title{Mutual Dependence Measures}
\usage{
mdm(X, dim_comp = NULL, dist_comp = FALSE, type = c("complete"))
}
\arguments{
\item{X}{A matrix.}

\item{dim_comp}{The dimensions of components in X \code{X}.}

\item{dist_comp}{Logical. If \code{TRUE}, the distances between all components
in X will be returned.}

\item{type}{The type of mutual dependence measures, including
- complete measure based on complete V-statistics
- simplified complete measure based on incomplete V-statistics
- asymmetric measure based on distance covariance
- symmectric measure based on distance covariance}
}
\value{
\code{mdm} returns a list containing the following components:
\item{stat}{The value of mutual dependence measure.}
\item{dist}{The distances between all components.}
}
\description{
\code{mdm} computes mutual dependence measures.
}
\examples{
X <- matrix(rnorm(30), 10, 3)
mdm(X)
}
\references{
Jin, Z. and Matteson, D. S. (2017).
  Generalizing Distance Covariance to Measure and Test Multivariate Mutual Dependence.
  arXiv preprint arXiv:1709.02532.
  \url{https://arxiv.org/abs/1709.02532}.
}
