% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDMeasure-package.R
\docType{package}
\name{MDMeasure-package}
\alias{MDMeasure-package}
\alias{MDMeasure}
\title{Mutual Dependence Measures via Energy Statistics}
\description{
MDMeasure: A package for mutual dependence measures via energy statistics
}
\details{
The MDMeasure package provides measures of mutual dependence and tests of mutual independence.
}
\section{Measuring mutual dependence}{

The mutual dependence measures:
- complete measure based on complete V-statistics
- simplified complete measure based on incomplete V-statistics
- asymmetric measure based on distance covariance
- symmectric measure based on distance covariance
}

\section{Initializing local optimization methods}{

The permutation tests based on the mutual dependence measures.
}

\author{
Ze Jin \email{zj58@cornell.edu}
}
