% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDMeasure-package.R
\docType{package}
\name{MDMeasure-package}
\alias{MDMeasure-package}
\alias{MDMeasure}
\title{Mutual Dependence Measures via Energy Statistics}
\description{
MDMeasure: A package for mutual dependence measures via energy statistics
}
\details{
The MDMeasure package provides measures of mutual dependence and tests of mutual independence.
}
\author{
Ze Jin \email{zj58@cornell.edu}
}
