% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDMeasure-package.R
\docType{package}
\name{MDMeasure-package}
\alias{MDMeasure-package}
\alias{MDMeasure}
\title{Mutual Dependence Measures via Energy Statistics}
\description{
MDMeasure: A package for mutual dependence measures via energy statistics
}
\details{
The MDMeasure package provides measures of mutual dependence and tests of mutual independence.
}
\section{Measuring mutual dependence}{


The mutual dependence measures include:

- asymmetric measure \eqn{\mathcal{R}_n} based on distance covariance \eqn{\mathcal{V}_n}

- symmectric measure \eqn{\mathcal{S}_n} based on distance covariance \eqn{\mathcal{V}_n}

- complete measure \eqn{\mathcal{Q}_n} based on complete V-statistics

- simplified complete measure \eqn{\mathcal{Q}_n^\star} based on incomplete V-statistics

- asymmetric measure \eqn{\mathcal{J}_n} based on complete measure \eqn{\mathcal{Q}_n}

- simplified asymmetric measure \eqn{\mathcal{J}_n^\star} based on simplified complete measure 
  \eqn{\mathcal{Q}_n^\star}

- symmectric measure \eqn{\mathcal{I}_n} based on complete measure \eqn{\mathcal{Q}_n}

- simplified symmectric measure \eqn{\mathcal{I}_n^\star} based on simplified complete measure 
  \eqn{\mathcal{Q}_n^\star}
}

\section{Testing mutual independence}{


The mutual independence tests based on the mutual dependence measures are implemented as permutation tests.
}

\author{
Ze Jin \email{zj58@cornell.edu}
}
